third_party/blink/web_tests/platform/win/external/wpt/css/css-writing-modes/forms/textarea-rows-cols-sizing-expected.txt

This is a testharness.js-based test.
PASS textarea[style="writing-mode: horizontal-tb"] rows attribute changes the size correctly
PASS textarea[style="writing-mode: horizontal-tb"] cols attribute changes the size correctly
FAIL textarea[style="writing-mode: vertical-lr"] rows attribute changes the size correctly assert_equals: height shouldn't change when changing rows expected 38 but got 166
FAIL textarea[style="writing-mode: vertical-lr"] cols attribute changes the size correctly assert_equals: width shouldn't change when changing cols expected 181 but got 341
FAIL textarea[style="writing-mode: vertical-rl"] rows attribute changes the size correctly assert_equals: height shouldn't change when changing rows expected 38 but got 166
FAIL textarea[style="writing-mode: vertical-rl"] cols attribute changes the size correctly assert_equals: width shouldn't change when changing cols expected 181 but got 341
Harness: the test ran to completion.
